complete split of async_request.py from channel.py
