+++ src/amuse/datamodel/memory_storage.py
@@ -8,6 +8,7 @@ from amuse.units.quantities import VectorQuantity
 from amuse.units.quantities import is_quantity
 from amuse.datamodel.base import *
 from amuse.support import exceptions
+from amuse.rfi.async_request import FakeASyncRequest
 
 try:
   if numpy.uintp().itemsize<8: raise Exception()
@@ -110,7 +111,6 @@ class InMemoryAttributeStorage(AttributeStorage):
         return results
     
     def get_values_in_store_async(self, indices, attributes):
-        from amuse.rfi.channel import FakeASyncRequest
         result = self.get_values_in_store(indices, attributes)
         return FakeASyncRequest(result)
         
@@ -146,7 +146,6 @@ class InMemoryAttributeStorage(AttributeStorage):
                     raise AttributeError("exception in setting attribute '{0}', error was '{1}'".format(attribute, ex))
 
     def set_values_in_store_async(self, indices, attributes, list_of_values_to_set):
-        from amuse.rfi.channel import FakeASyncRequest
         result = self.set_values_in_store(indices, attributes, list_of_values_to_set)
         return FakeASyncRequest(result)
