+++ test/compile_tests/test_python_sockets_implementation.py
@@ -12,7 +12,7 @@ from amuse.units import units
 from amuse import datamodel
 from amuse.rfi import python_code
 from amuse.rfi.core import *
-from amuse.rfi.channel import AsyncRequestsPool
+from amuse.rfi.async_request import AsyncRequestsPool
 
             
 class ForTestingInterface(PythonCodeInterface):
